Separate exception handler
- Separate exception handler
- Addapt reserve to use room name and address.
